+++ test/ecto/integration/crud_test.exs
@@ -176,14 +176,9 @@ defmodule Ecto.Integration.CrudTest do
                TestRepo.all(Account, prefix: tenant)
                |> Enum.sort(fn %Account{name: a}, %Account{name: b} -> a <= b end)
 
-      assert_raise(
-        Unsupported,
-        ~r/the ordering must correspond to the primary key or an indexed field/,
-        fn ->
-          from(Account, order_by: :name)
-          |> TestRepo.all(prefix: tenant)
-        end
-      )
+      assert [%{name: "Jane" <> _}, %{name: "John" <> _}] =
+               from(Account, order_by: :name)
+               |> TestRepo.all(prefix: tenant)
     end
 
     test "tx_insert", context do