+++ test/ecto/integration/ordering_test.exs
@@ -109,14 +109,9 @@ defmodule Ecto.Integration.OrderingTest.OrderByDataField do
     tenant = context[:tenant]
     put_pk_data(tenant)
 
-    # order by non-pk, non-index field is not supported
-    assert_raise Unsupported, ~r/order_by/, fn ->
-      TestRepo.all(from(q in QueueItem, order_by: {:asc, :data}), prefix: tenant)
-    end
-
-    assert_raise Unsupported, ~r/order_by/, fn ->
-      TestRepo.all(from(q in QueueItem, order_by: {:desc, :data}), prefix: tenant)
-    end
+    # order by non-pk, non-index field is supported when there's no where clause
+    assert [_ | _] = TestRepo.all(from(q in QueueItem, order_by: {:asc, :data}), prefix: tenant)
+    assert [_ | _] = TestRepo.all(from(q in QueueItem, order_by: {:desc, :data}), prefix: tenant)
   end
 
   test "data field forward limit", context do
@@ -125,7 +120,7 @@ defmodule Ecto.Integration.OrderingTest.OrderByDataField do
 
     # limited ordering not possible, raise error instead of allowing this
     assert_raise Unsupported,
-                 ~r/the ordering must correspond to the primary key or an indexed field/,
+                 ~r/field must be part of the where constraint/,
                  fn ->
                    TestRepo.all(from(q in QueueItem, order_by: {:asc, :data}),
                      prefix: tenant,
@@ -149,7 +144,7 @@ defmodule Ecto.Integration.OrderingTest.OrderByDataField do
 
     # limited ordering not possible, raise error instead of allowing this
     assert_raise Unsupported,
-                 ~r/the ordering must correspond to the primary key or an indexed field/,
+                 ~r/field must be part of the where constraint/,
                  fn ->
                    TestRepo.all(from(q in QueueItem, order_by: {:asc, :data}, limit: 2),
                      prefix: tenant
@@ -180,14 +175,24 @@ defmodule Ecto.Integration.OrderingTest.IndexWithNoneConstraint do
     # empty because Event has `primary_write: false`
     assert [] = TestRepo.all(Event, prefix: tenant)
 
-    # index selection is not activated by order_by, where clause is required
-    assert_raise Unsupported, ~r//, fn ->
-      TestRepo.all(from(e in Event, order_by: [asc: e.date]), prefix: tenant)
-    end
+    all_events =
+      TestRepo.all(
+        from(e in Event, where: e.date > ^~D[0000-01-01] and e.date < ^~D[9999-01-01]),
+        prefix: tenant
+      )
 
-    assert_raise Unsupported, ~r/order_by/, fn ->
-      TestRepo.all(from(e in Event, order_by: [desc: e.date]), prefix: tenant)
-    end
+    assert 27 == length(all_events)
+
+    # index selection is activated by order_by
+    assert [%{date: ~D[2000-01-01]} | rest_events] =
+             TestRepo.all(from(e in Event, order_by: [asc: e.date]), prefix: tenant)
+
+    assert 26 == length(rest_events)
+
+    assert [%{date: ~D[2002-01-01]} | rest_events] =
+             TestRepo.all(from(e in Event, order_by: [desc: e.date]), prefix: tenant)
+
+    assert 26 == length(rest_events)
   end
 
   test "order by first field in index, with limit", context do
@@ -250,7 +255,7 @@ defmodule Ecto.Integration.OrderingTest.IndexWithNoneConstraint do
     put_idx_data(tenant)
 
     assert_raise Unsupported,
-                 ~r/When querying with an order_by, the ordering must correspond to the primary key or an indexed field/,
+                 ~r/field must be part of the where constraint/,
                  fn ->
                    TestRepo.all(from(e in Event, order_by: [asc: e.user_id]),
                      prefix: tenant,
@@ -265,7 +270,7 @@ defmodule Ecto.Integration.OrderingTest.IndexWithNoneConstraint do
     put_idx_data(tenant)
 
     assert_raise Unsupported,
-                 ~r/When querying with an order_by, the ordering must correspond to the primary key or an indexed field/,
+                 ~r/field must be part of the where constraint/,
                  fn ->
                    TestRepo.all(from(e in Event, order_by: [asc: e.user_id], limit: 1),
                      prefix: tenant
@@ -458,5 +463,35 @@ defmodule Ecto.Integration.OrderingTest.QueryLimitWithSplitObjects do
 
     assert [%{id: "alice"}, %{id: "bob"}] =
              TestRepo.all(from(u in User, limit: 2), prefix: tenant, key_limit: 4)
+  end
+end
+
+defmodule Ecto.Integration.OrderingTest.MultipleOrderBys do
+  use Ecto.Integration.Case, async: true
+
+  alias EctoFoundationDB.Schemas.User
+  alias EctoFoundationDB.Test.Util
+
+  import Ecto.Query
+
+  test "option limit", context do
+    tenant = context[:tenant]
+
+    users = [
+      %User{id: "1", name: "Alice", notes: "foo", inserted_at: ~N[2026-02-01 00:00:01]},
+      %User{id: "2", name: "Bob", notes: "foo", inserted_at: ~N[2026-02-01 00:00:02]},
+      %User{id: "3", name: "Alice", notes: "bar", inserted_at: ~N[2026-02-01 00:00:03]},
+      %User{id: "4", name: "Alice", notes: "foo", inserted_at: ~N[2026-02-01 00:00:04]},
+      %User{id: "5", name: "Bob", notes: "foo", inserted_at: ~N[2026-02-01 00:00:05]},
+      %User{id: "6", name: "Alice", notes: "bar", inserted_at: ~N[2026-02-01 00:00:06]}
+    ]
+
+    for u <- users, do: TestRepo.insert(u, prefix: tenant)
+
+    query =
+      from(u in User, order_by: [asc: u.name, desc: u.inserted_at])
+
+    assert [%{id: "6"}, %{id: "4"}, %{id: "3"}, %{id: "1"}, %{id: "5"}, %{id: "2"}] =
+             TestRepo.all(query, prefix: tenant)
   end
 end