+++ test/ecto_foundationdb/layer/metadata_test.exs
@@ -11,40 +11,40 @@ defmodule EctoFoundationDBLayerMetadataTest do
       idx_a = [id: :a, fields: [:user_id]]
       idx_b = [id: :b, fields: [:user_id]]
       constraints = [%Equal{field: :user_id}]
-      assert ^idx_a = Metadata.select_index(Metadata.new([idx_a, idx_b]), constraints)
-      assert ^idx_b = Metadata.select_index(Metadata.new([idx_b, idx_a]), constraints)
+      assert ^idx_a = Metadata.select_index(Metadata.new([idx_a, idx_b]), constraints, [])
+      assert ^idx_b = Metadata.select_index(Metadata.new([idx_b, idx_a]), constraints, [])
     end
 
     test "b over a" do
       idx_a = [id: :a, fields: [:timestamp]]
       idx_b = [id: :b, fields: [:user_id]]
       constraints = [%Equal{field: :user_id}]
-      assert ^idx_b = Metadata.select_index(Metadata.new([idx_a, idx_b]), constraints)
-      assert ^idx_b = Metadata.select_index(Metadata.new([idx_b, idx_a]), constraints)
+      assert ^idx_b = Metadata.select_index(Metadata.new([idx_a, idx_b]), constraints, [])
+      assert ^idx_b = Metadata.select_index(Metadata.new([idx_b, idx_a]), constraints, [])
     end
 
     test "exact matches with different between ordering" do
       idx_a = [id: :a, fields: [:date, :time, :user_id]]
       idx_b = [id: :b, fields: [:date, :user_id, :time]]
       constraints = [%Equal{field: :date}, %Equal{field: :user_id}, %Between{field: :time}]
-      assert ^idx_b = Metadata.select_index(Metadata.new([idx_a, idx_b]), constraints)
-      assert ^idx_b = Metadata.select_index(Metadata.new([idx_b, idx_a]), constraints)
+      assert ^idx_b = Metadata.select_index(Metadata.new([idx_a, idx_b]), constraints, [])
+      assert ^idx_b = Metadata.select_index(Metadata.new([idx_b, idx_a]), constraints, [])
     end
 
     test "one subset, the other insufficient" do
       idx_a = [id: :a, fields: [:date, :time, :user_id]]
       idx_b = [id: :b, fields: [:user_id, :time]]
       constraints = [%Equal{field: :date}]
-      assert ^idx_a = Metadata.select_index(Metadata.new([idx_a, idx_b]), constraints)
-      assert ^idx_a = Metadata.select_index(Metadata.new([idx_b, idx_a]), constraints)
+      assert ^idx_a = Metadata.select_index(Metadata.new([idx_a, idx_b]), constraints, [])
+      assert ^idx_a = Metadata.select_index(Metadata.new([idx_b, idx_a]), constraints, [])
     end
 
     test "inexact matches with different between ordering" do
       idx_a = [id: :a, fields: [:date, :time, :user_id, :extra]]
       idx_b = [id: :b, fields: [:date, :user_id, :time, :extra]]
       constraints = [%Equal{field: :date}, %Equal{field: :user_id}, %Between{field: :time}]
-      assert ^idx_b = Metadata.select_index(Metadata.new([idx_a, idx_b]), constraints)
-      assert ^idx_b = Metadata.select_index(Metadata.new([idx_b, idx_a]), constraints)
+      assert ^idx_b = Metadata.select_index(Metadata.new([idx_a, idx_b]), constraints, [])
+      assert ^idx_b = Metadata.select_index(Metadata.new([idx_b, idx_a]), constraints, [])
     end
 
     test "best partial match" do
@@ -52,8 +52,10 @@ defmodule EctoFoundationDBLayerMetadataTest do
       idx_a = [id: :a, fields: [:date, :time]]
       idx_b = [id: :b, fields: [:date, :user_id, :time]]
       constraints = [%Between{field: :time}]
-      assert ^idx_a = Metadata.select_index(Metadata.new([idx_a, idx_b]), constraints)
-      assert ^idx_a = Metadata.select_index(Metadata.new([idx_b, idx_a]), constraints)
+      assert ^idx_a = Metadata.select_index(Metadata.new([idx_a, idx_b]), constraints, [])
+      assert ^idx_a = Metadata.select_index(Metadata.new([idx_b, idx_a]), constraints, [])
     end
   end
+
+  # @todo: Add tests for ordering behaviors
 end